misc:  Clean up some of the problems compiling for Windows with mingw
Major remaining tasks:

  * Forcing Non-blocking I/O
  * dynamic library loading (eg dlopen or ltdl)
  * Final linking

+++ backend_canonselphy.c
@@ -457,9 +457,9 @@ enum {
 	S_FINISHED,
 };
 
-static int fancy_memcmp(const uint8_t *buf_a, const int16_t *buf_b, uint len)
+static int fancy_memcmp(const uint8_t *buf_a, const int16_t *buf_b, uint16_t len)
 {
-	uint i;
+	uint16_t i;
 
 	for (i = 0 ; i < len ; i++) {
 		if (buf_b[i] == -1)
